feat: move mock to client-api

+++ packages/client-api/src/requests/device.ts
@@ -1,9 +1,13 @@
 import { AxiosInstance } from 'axios';
 import { IDevice, IResponse } from '@lib/types';
+import { device as mockDevice, devices as mockDevices } from '@lib/mock';
 
+import { sleep } from '../utils';
 import { BaseApi } from '../requests/baseApi';
-
 import { error, device as types } from '../types';
+
+const isMock = process.env.MOCK;
+const mockTimeout = 500;
 
 class Device extends BaseApi {
   constructor(api: AxiosInstance, deviceId: string) {
@@ -11,10 +15,20 @@ class Device extends BaseApi {
   }
 
   addDevice = async (deviceName: string, userId: string) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'ADD_DEVICE',
+        id: mockDevice.id,
+      } as types.IAddDeviceResponse;
+    }
+
     const body = {
       deviceName,
       userId,
     };
+
     const res = await this.api.post<IResponse<string>>('/devices', body);
     const resData = res.data;
 
@@ -38,6 +52,15 @@ class Device extends BaseApi {
     * @returns
     */
   getDevices = async (userId?: string) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'GET_DEVICES',
+        devices: mockDevices,
+      } as types.IGetDevicesResponse;
+    }
+
     const res = await this.api.get<IResponse<IDevice[]>>(`/devices?${userId ? `userId=${userId}` : ''}`);
     const resData = res.data;
 
@@ -62,6 +85,15 @@ class Device extends BaseApi {
     * @returns IDevice
     */
   getDevice = async (deviceId: string, userId?: string) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'GET_DEVICE',
+        device: mockDevice,
+      } as types.IGetDeviceResponse;
+    }
+
     try {
       const paramQuery = userId ? `?userId=${userId}` : '';
 
@@ -105,6 +137,15 @@ class Device extends BaseApi {
   };
 
   updateDevice = async (device: Partial<IDevice>) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'UPDATE_DEVICE',
+        deviceId: device.id,
+      } as types.IUpdateDeviceResponse;
+    }
+
     const res = await this.api.patch<IResponse<string>>(`/devices/${device.id}`, device);
     const resData = res.data;
 
@@ -121,6 +162,14 @@ class Device extends BaseApi {
   };
 
   removeDevice = async (deviceId: string) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'REMOVE_DEVICE',
+      } as types.IRemoveDeviceResponse;
+    }
+
     const res = await this.api.delete<IResponse<void>>(`/devices/${deviceId}`);
     const resData = res.data;
